Restructure GraphQL folder
