+++ GraphQL/main.py
@@ -1,5 +1,5 @@
 import uvicorn
-from app import app
+from src.app import app
 
 if __name__ == "__main__":
     uvicorn.run(app, host="0.0.0.0", port=5001)
