+++ GraphQL/src/app.py
@@ -1,5 +1,5 @@
 import strawberry.asgi
-from schema import Query, Mutation
+from src.schema import Query, Mutation
 
 schema = strawberry.Schema(query=Query, mutation=Mutation)
